fix JSON encoding for no tokens (empty array);
