+++ ergotree-ir/src/chain/ergo_box.rs
@@ -58,33 +58,22 @@ pub type BoxTokens = BoundedVec<Token, 1, 255>;
 )]
 #[derive(PartialEq, Eq, Debug, Clone)]
 pub struct ErgoBox {
-    #[cfg_attr(feature = "json", serde(rename = "boxId", alias = "id"))]
-    box_id: BoxId,
+    pub(crate) box_id: BoxId,
     /// amount of money associated with the box
-    #[cfg_attr(feature = "json", serde(rename = "value"))]
     pub value: BoxValue,
     /// guarding script, which should be evaluated to true in order to open this box
-    #[cfg_attr(
-        feature = "json",
-        serde(rename = "ergoTree", with = "crate::chain::json::ergo_tree")
-    )]
     pub ergo_tree: ErgoTree,
     /// secondary tokens the box contains
-    #[cfg_attr(feature = "json", serde(rename = "assets"))]
     pub tokens: Option<BoxTokens>,
     ///  additional registers the box can carry over
-    #[cfg_attr(feature = "json", serde(rename = "additionalRegisters"))]
     pub additional_registers: NonMandatoryRegisters,
     /// height when a transaction containing the box was created.
     /// This height is declared by user and should not exceed height of the block,
     /// containing the transaction with this box.
-    #[cfg_attr(feature = "json", serde(rename = "creationHeight"))]
     pub creation_height: u32,
     /// id of transaction which created the box
-    #[cfg_attr(feature = "json", serde(rename = "transactionId", alias = "txId"))]
     pub transaction_id: TxId,
     /// number of box (from 0 to total number of boxes the transaction with transactionId created - 1)
-    #[cfg_attr(feature = "json", serde(rename = "index"))]
     pub index: u16,
 }
 
@@ -224,93 +213,28 @@ impl SigmaSerializable for ErgoBox {
     }
 }
 
-#[cfg(feature = "json")]
-impl TryFrom<super::json::ergo_box::ErgoBoxJson> for ErgoBox {
-    type Error = super::json::ergo_box::ErgoBoxFromJsonError;
-    fn try_from(box_json: super::json::ergo_box::ErgoBoxJson) -> Result<Self, Self::Error> {
-        let tokens = if box_json.tokens.is_empty() {
-            None
-        } else {
-            Some(box_json.tokens.try_into().map_err(|_| {
-                SigmaSerializationError::NotSupported(
-                    "More than 255 tokens are not allowed in a box",
-                )
-            })?)
-        };
-        let box_with_zero_id = ErgoBox {
-            box_id: BoxId::zero(),
-            value: box_json.value,
-            ergo_tree: box_json.ergo_tree,
-            tokens,
-            additional_registers: box_json.additional_registers,
-            creation_height: box_json.creation_height,
-            transaction_id: box_json.transaction_id,
-            index: box_json.index,
-        };
-        let box_id = box_with_zero_id.calc_box_id()?;
-        let ergo_box = ErgoBox {
-            box_id,
-            ..box_with_zero_id
-        };
-        match box_json.box_id {
-            Some(box_id) => {
-                if ergo_box.box_id() == box_id {
-                    Ok(ergo_box)
-                } else {
-                    Err(super::json::ergo_box::ErgoBoxFromJsonError::InvalidBoxId)
-                }
-            }
-            None => Ok(ergo_box),
-        }
-    }
-}
-
-#[cfg(feature = "json")]
-impl From<ErgoBox> for super::json::ergo_box::ErgoBoxJson {
-    fn from(ergo_box: ErgoBox) -> super::json::ergo_box::ErgoBoxJson {
-        let tokens = ergo_box
-            .tokens
-            .as_ref()
-            .map(BoxTokens::as_vec)
-            .cloned()
-            .unwrap_or_default(); // JSON serialization for ErgoBox requires that tokens be [] instead of null
-        super::json::ergo_box::ErgoBoxJson {
-            box_id: Some(ergo_box.box_id),
-            value: ergo_box.value,
-            ergo_tree: ergo_box.ergo_tree,
-            tokens,
-            additional_registers: ergo_box.additional_registers,
-            creation_height: ergo_box.creation_height,
-            transaction_id: ergo_box.transaction_id,
-            index: ergo_box.index,
-        }
-    }
-}
 /// Contains the same fields as `ErgoBox`, except for transaction id and index,
 /// that will be calculated after full transaction formation.
 /// Use `ErgoBoxCandidateBuilder` from ergo-lib crate to create an instance.
 #[cfg_attr(feature = "json", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(
+    feature = "json",
+    serde(try_from = "crate::chain::json::ergo_box::ErgoBoxCandidateJson"),
+    serde(into = "crate::chain::json::ergo_box::ErgoBoxCandidateJson")
+)]
 #[derive(PartialEq, Eq, Clone, Debug)]
 pub struct ErgoBoxCandidate {
     /// amount of money associated with the box
-    #[cfg_attr(feature = "json", serde(rename = "value"))]
     pub value: BoxValue,
     /// guarding script, which should be evaluated to true in order to open this box
-    #[cfg_attr(
-        feature = "json",
-        serde(rename = "ergoTree", with = "crate::chain::json::ergo_tree")
-    )]
     pub ergo_tree: ErgoTree,
     /// secondary tokens the box contains
-    #[cfg_attr(feature = "json", serde(rename = "assets"))]
     pub tokens: Option<BoxTokens>,
     ///  additional registers the box can carry over
-    #[cfg_attr(feature = "json", serde(rename = "additionalRegisters"))]
     pub additional_registers: NonMandatoryRegisters,
     /// height when a transaction containing the box was created.
     /// This height is declared by user and should not exceed height of the block,
     /// containing the transaction with this box.
-    #[cfg_attr(feature = "json", serde(rename = "creationHeight"))]
     pub creation_height: u32,
 }
